requirements.txt
absl-py==1.0.0
alabaster==0.7.12
appnope==0.1.2
astunparse==1.6.3
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.10.0
blis==0.2.4
bpemb==0.3.3
cachetools==5.0.0
catalogue==2.0.6
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
cloudpickle==1.6.0
commonmark==0.9.1
cycler==0.11.0
cymem==2.0.6
Cython==0.29.26
dataclasses==0.6
decorator==5.1.1
Deprecated==1.2.13
docutils==0.15.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.1.0/en_core_web_sm-2.1.0.tar.gz
filelock==3.4.2
flair==0.6.1.post1
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
ftfy==6.0.3
future==0.18.2
fvcore==0.1.5.post20220119
gast==0.3.3
gdown==4.2.0
gensim==4.1.2
google-auth==2.5.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.43.0
h5py==2.10.0
huggingface-hub==0.4.0
hyperopt==0.2.7
idna==3.3
imageio==2.14.1
imagesize==1.3.0
importlib-metadata==3.10.1
iopath==0.1.9
ipython==7.23.1
itsdangerous==2.0.1
Janome==0.4.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
konoha==4.6.5
langdetect==1.0.9
lxml==4.7.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.3.1
matplotlib-inline==0.1.3
mpld3==0.3
murmurhash==1.0.6
mysql-connector-python==8.0.21
networkx==2.6.3
nltk==3.5
numpy<=1.22.1
oauthlib==3.1.1
opencv-python==4.4.0.42
opt-einsum==3.3.0
overrides==3.1.0
packaging==21.3
pandas==1.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
plac==0.9.6
portalocker==2.3.2
preshed==2.0.1
prompt-toolkit==3.0.26
protobuf==3.19.3
ptyprocess==0.7.0
py4j==0.10.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pydantic==1.8.2
Pygments==2.11.2
PyMySQL==1.0.2
pyparsing==3.0.7
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.2.0
PyYAML==5.4.1
recommonmark==0.7.1
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
sacremoses==0.0.47
scikit-image==0.19.1
scikit-learn==0.24.2
scipy==1.4.1
seaborn==0.11.0
segtok==1.5.11
sentencepiece==0.1.96
six==1.16.0
smart-open==5.2.1
snowballstemmer==2.2.0
soupsieve==2.3.1
spacy==2.1.9
Sphinx==4.3.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.31
sqlitedict==1.7.0
srsly==1.0.5
tabulate==0.8.7
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.3.0
tensorflow-estimator==2.3.0
tensorflow-hub==0.10.0
termcolor==1.1.0
thinc==7.0.8
threadpoolctl==3.0.0
tifffile==2021.11.2
tokenizers==0.11.4
torch==1.8.1
torchtext==0.4.0
torchvision==0.9.1
tqdm==4.30.0
traitlets==5.1.1
transformers==4.16.0
typing_extensions==4.0.1
urllib3==1.26.8
wasabi==0.9.0
wcwidth==0.2.5
Werkzeug==2.0.2
wordcloud==1.8.1
wrapt==1.13.3
yacs==0.1.8
zipp==3.7.0
